Move substitution to separate package
